backend/requirements.txt
crewai>=0.28.0
groq>=0.4.0
requests>=2.31.0
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.0.0
python-dotenv>=1.0.0
reportlab>=4.0.0
matplotlib>=3.8.0
Pillow>=10.0.0
pymongo>=4.6.0
dnspython>=2.4.0
PyJWT[crypto]>=2.8.0
PyPDF2>=3.0.0
python-pptx>=0.6.21
openpyxl>=3.1.0
python-docx>=1.0.0
python-multipart>=0.0.6
beautifulsoup4>=4.12.0
openai==1.83.0
litellm==1.56.2
fastapi-sso>=0.20.0
apscheduler>=3.11.0
SpeechRecognition>=3.10.0
playwright>=1.40.0
langchain_groq>=0.1.0
certifi>=2023.7.22
email-validator>=2.0.0